add runit.sh

+++ .bin/fonts.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 set -eu
 
 if [ "$(uname -s)" = "Linux" ]; then
@@ -17,8 +18,10 @@ wget https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20B
 wget https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Italic.ttf -P "$FONTS_DIR/MesloLGS_NF" >/dev/null 2>&1
 wget https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Bold%20Italic.ttf -P "$FONTS_DIR/MesloLGS_NF" >/dev/null 2>&1
 
-CICA_RELEASES_URL="https://api.github.com/repos/miiton/Cica/releases"
-curl -sfL "${CICA_RELEASES_URL}" | jq -r '.[0].assets | .[].browser_download_url' | grep -i without_emoji.zip | xargs -I{} curl -fL -o ~/Cica.zip "{}"
-(cd ~/ && unzip -o Cica.zip -d "$FONTS_DIR/cica")
+if [ "$(uname -s)" = "Linux" ]; then
+    CICA_RELEASES_URL="https://api.github.com/repos/miiton/Cica/releases"
+    curl -sfL "${CICA_RELEASES_URL}" | jq -r '.[0].assets | .[].browser_download_url' | grep -i without_emoji.zip | xargs -I{} curl -fL -o ~/Cica.zip "{}"
+    (cd ~/ && unzip -o Cica.zip -d "$FONTS_DIR/cica")
 
-rm -f ~/Cica.zip
+    rm -f ~/Cica.zip
+fi
